поменять название классов, добавить правило для линтера import/no-unresolved
