+++ package.json
@@ -15,7 +15,7 @@
     "react-scripts": "5.0.1",
     "redux": "^4.2.1",
     "styled-components": "^6.1.0",
-    "swiper": "^10.3.1",
+    "swiper": "^11.0.3",
     "web-vitals": "^2.1.0"
   },
   "scripts": {
@@ -42,6 +42,12 @@
     "@types/redux": "^3.6.0",
     "@types/styled-components": "^5.1.28",
     "@types/swiper": "^6.0.0",
+    "@typescript-eslint/eslint-plugin": "^6.9.0",
+    "@typescript-eslint/parser": "^6.9.0",
+    "eslint": "^8.50.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-import": "^2.28.1",
     "gh-pages": "^6.0.0",
     "prettier": "^3.0.3",
     "typescript": "^5.2.2"
