+++ src/components/cart/cart.module.css
@@ -1,4 +1,4 @@
-.cart {
+.container {
   background-color: #fff;
   box-shadow: 0 4px 28px rgba(0, 0, 0, 0.2);
   border-radius: 10px;
@@ -10,7 +10,7 @@
   position: absolute;
 }
 
-.cart::before {
+.container::before {
   top: -12px;
   right: 50px;
   position: absolute;
@@ -22,33 +22,33 @@
   background: #fff;
 }
 
-.cart__list {
+.list {
   padding: 0;
   margin: 0;
   list-style: none;
 }
 
-.cart__item {
+.item {
   padding: 24px;
   display: flex;
 }
 
-.cart__item-image {
+.itemImage {
   display: block;
   width: 71px;
   height: 71px;
 }
 
-.cart__item-body {
+.itemBody {
   margin-left: 17px;
 }
 
-.cart__item-top {
+.itemTop {
   display: flex;
   justify-content: space-between;
 }
 
-.cart__item-title {
+.itemTitle {
   color: #0e0c0d;
   font-size: 15px;
   font-style: normal;
@@ -56,7 +56,7 @@
   line-height: 17px;
 }
 
-.cart__trash-button {
+.trashButton {
   width: 20px;
   height: 23px;
   border: none;
@@ -65,58 +65,58 @@
   background-repeat: no-repeat;
 }
 
-.cart__item-bottom {
+.itemBottom {
   display: flex;
   justify-content: space-between;
 }
 
-.cart__item-buttons {
+.itemButtons {
 }
 
-.cart__button-increase {
+.buttonIncrease {
 }
 
-.cart__item-counter {
+.itemCounter {
 }
 
-.cart__item-decrease {
+.itemDecrease {
 }
 
-.cart__item-total {
+.itemTotal {
 }
 
-.cart__footer {
+.footer {
 }
 
-.cart__footer-container {
+.footerContainer {
 }
 
-.cart__total {
+.total {
 }
 
-.cart__total-price {
+.totalPrice {
 }
 
-.cart__price {
+.price {
 }
 
-.cart__add-to-order {
+.addToOrder {
 }
 
-.cart__extra-list {
+.extraList {
 }
 
-.cart__extra-item {
+.extraItem {
 }
 
-.cart__extra-item-image {
+.extraItemImage {
 }
 
-.cart__extra-item-body {
+.extraItemBody {
 }
 
-.cart__extra-item-title {
+.extraItemTitle {
 }
 
-.cart__extra-item-text {
+.extraItemText {
 }
