+++ src/components/feedback/feedback.module.css
@@ -1,9 +1,9 @@
-.feedback {
+.container {
   width: 100%;
   padding-top: 50px;
 }
 
-.feedback__heading {
+.title {
   color: #231f20;
   font-family: Montserrat, serif;
   font-size: 24px;
@@ -12,7 +12,7 @@
   line-height: normal;
 }
 
-.feedback__list {
+.list {
   display: flex;
   justify-content: space-between;
   gap: 46px;
@@ -21,7 +21,7 @@
   margin: 0;
 }
 
-.feedback__item {
+.item {
   width: 97px;
   height: 98px;
   flex-shrink: 0;
@@ -31,18 +31,18 @@
   transition: 0.6s;
 }
 
-.feedback__item:hover {
+.item:hover {
   border-color: #ff2e65;
 }
 
-.feedback__image {
+.image {
   width: 100px;
   height: 100%;
   object-fit: cover;
   display: block;
 }
 
-.feedback__image-description {
+.imageDescription {
   text-align: center;
   margin: 17px 0 0;
   color: #9d9d9b;
@@ -53,11 +53,11 @@
   line-height: normal;
 }
 
-.feedback__link {
+.link {
   color: inherit;
   text-decoration: none;
 }
 
-.feedback_slide:hover .feedback__item {
+.slide:hover .item {
   border: 4px solid #dd154a;
 }
