+++ src/components/menu/menu.module.css
@@ -1,12 +1,8 @@
-.menu {
-
-}
-
-.menu__category {
+.category {
   padding-top: 55px;
 }
 
-.menu__heading {
+.title {
   padding: 0 0 22px;
   color: #ff2e65;
   font-size: 30px;
@@ -15,7 +11,7 @@
   line-height: normal;
 }
 
-.menu__list {
+.list {
   list-style: none;
   margin: 0;
   padding: 0;
@@ -24,7 +20,7 @@
   gap: 30px;
 }
 
-.menu__item-label {
+.itemLabel {
   top: 0;
   right: 0;
   position: absolute;
@@ -40,19 +36,19 @@
   box-shadow: 0 2px 5px rgba(0, 0, 0, 0.3);
 }
 
-.menu__item {
+.item {
   position: relative;
   display: flex;
   flex-direction: column;
   align-items: center;
 }
 
-.menu__item-image {
+.itemImage {
   width: 100%;
   display: block;
 }
 
-.menu__item-heading {
+.itemTitle {
   padding: 0;
   margin: 4px 0 0;
   color: #0e0c0d;
@@ -62,7 +58,7 @@
   line-height: 28px;
 }
 
-.menu__item-description {
+.itemDescription {
   padding: 0;
   margin: 18px 0 auto;
   color: #686466;
@@ -72,14 +68,14 @@
   line-height: 19px;
 }
 
-.menu__item-footer {
+.itemFooter {
   width: 100%;
   margin: 23px 0 0;
   display: flex;
   justify-content: space-between;
 }
 
-.menu__item-price {
+.itemPrice {
   padding: 0;
   margin: 0;
   color: #231f20;
@@ -89,7 +85,7 @@
   line-height: 19px;
 }
 
-.menu__button {
+.button {
   padding: 4px 24px;
   border: none;
   border-radius: 8px;
@@ -99,9 +95,9 @@
   font-style: normal;
   font-weight: 700;
   line-height: 28px;
-  transition: .6s;
+  transition: 0.6s;
 }
 
-.menu__button:hover {
+.button:hover {
   background: #cc2551;
 }
