--- src/components/new-product-details/index.module.css
@@ -1,33 +1,0 @@
-.new-product-details {
-    display: flex;
-    align-items: center;
-}
-
-.new-product-details__heading {
-    color: #0E0C0D;
-    font-size: 28px;
-    font-style: normal;
-    font-weight: 800;
-    line-height: 28px; /* 116.667% */
-    padding: 10px 20px;
-    margin: 0;
-}
-
-.new-product-details__description {
-    padding: 0;
-    margin: 0;
-    color: #686466;
-    font-size: 20px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 19px;
-}
-
-.new-product-details__button {
-    background-color: #ff2e65;
-    border: none;
-}
-
-.new-product-details__button:hover {
-    background: #cc2551;
-}
