+++ src/components/new-product-details/new-product-details.module.css
@@ -1,0 +1,33 @@
+.container {
+  display: flex;
+  align-items: center;
+}
+
+.title {
+  color: #0e0c0d;
+  font-size: 28px;
+  font-style: normal;
+  font-weight: 800;
+  line-height: 28px;
+  padding: 10px 20px;
+  margin: 0;
+}
+
+.description {
+  padding: 0;
+  margin: 0;
+  color: #686466;
+  font-size: 20px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 19px;
+}
+
+.button {
+  background-color: #ff2e65;
+  border: none;
+}
+
+.button:hover {
+  background: #cc2551;
+}
