--- src/components/new-products/index.module.css
@@ -1,86 +1,0 @@
-.new-products {
-    padding: 50px 0 0;
-    position: relative;
-}
-
-.new-products__heading {
-    padding-bottom: 33px;
-    margin: 0;
-    color: #231F20;
-    font-size: 24px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: normal;
-}
-
-.new-products__list {
-    margin: 0;
-    padding: 0;
-    list-style: none;
-    display: flex;
-    justify-content: space-between;
-}
-
-.new-products__list::before {
-    top: 60px;
-    left: -70px;
-    z-index: -1;
-    position: absolute;
-    content: url('../../images/onion.jpg');
-}
-
-.new-products__item {
-    text-decoration: none;
-    cursor: pointer;
-    width: 100%;
-    max-width: 255px;
-    padding: 14px 35px 14px 17px;
-    display: flex;
-    border-radius: 12px;
-    background: #FFF;
-    box-shadow: 0 4px 24px 0 rgba(0, 0, 0, 0.2);
-    transition: .6s;
-}
-
-.new-products__item:hover {
-    box-shadow: none;
-}
-
-.new-products__image {
-    margin-right: 25px;
-    display: block;
-    width: 71px;
-    height: 71px;
-}
-
-.new-products__item-wrapper {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.new-products__item-heading {
-    text-align: center;
-    margin: 0;
-    padding: 0;
-    color: #000;
-    font-size: 18px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: normal;
-}
-
-.new-products__item-price {
-    text-align: center;
-    margin: 0;
-    padding: 0;
-    color: #FF2E65;
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: normal;
-}
-
-
-
