+++ src/components/new-products/new-products.module.css
@@ -1,0 +1,83 @@
+.container {
+  padding: 50px 0 0;
+  position: relative;
+}
+
+.title {
+  padding-bottom: 33px;
+  margin: 0;
+  color: #231f20;
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: normal;
+}
+
+.list {
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  display: flex;
+  justify-content: space-between;
+}
+
+.list::before {
+  top: 60px;
+  left: -70px;
+  z-index: -1;
+  position: absolute;
+  content: url("../../images/onion.jpg");
+}
+
+.item {
+  text-decoration: none;
+  cursor: pointer;
+  width: 100%;
+  max-width: 255px;
+  padding: 14px 35px 14px 17px;
+  display: flex;
+  border-radius: 12px;
+  background: #fff;
+  box-shadow: 0 4px 24px 0 rgba(0, 0, 0, 0.2);
+  transition: 0.6s;
+}
+
+.item:hover {
+  box-shadow: none;
+}
+
+.image {
+  margin-right: 25px;
+  display: block;
+  width: 71px;
+  height: 71px;
+}
+
+.itemWrapper {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.itemTitle {
+  text-align: center;
+  margin: 0;
+  padding: 0;
+  color: #000;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: normal;
+}
+
+.itemPrice {
+  text-align: center;
+  margin: 0;
+  padding: 0;
+  color: #ff2e65;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: normal;
+}
