+++ src/components/pages/error-page/error-page.module.css
@@ -1,4 +1,4 @@
-.error {
+.container {
   height: 100vh;
   display: flex;
   flex-direction: column;
@@ -6,7 +6,7 @@
   align-items: center;
 }
 
-.error__title {
+.title {
   color: #0e0c0d;
   font-size: 60px;
   font-style: normal;
@@ -14,7 +14,7 @@
   line-height: 28px;
 }
 
-.error__image {
+.image {
   margin-top: 100px;
   display: block;
 }
