+++ src/components/slider/slider.module.css
@@ -1,7 +1,7 @@
-.wrapper {
+.container {
   margin-top: 28px;
 }
 
-.wrapper_fixed {
+.containerFixed {
   margin-top: 163px;
 }
