+++ src/index.tsx
@@ -1,16 +1,15 @@
-import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import "bootstrap/dist/css/bootstrap.min.css";
-import App from "./components/app";
 import { Provider } from "react-redux/es/exports";
-import store from "./services/store";
+import App from "./components/app/app.tsx";
+import store from "./services/store/store";
 
 const root = ReactDOM.createRoot(
-  document.getElementById("root") as HTMLElement
+  document.getElementById("root") as HTMLElement,
 );
 root.render(
   <Provider store={store}>
     <App />
-  </Provider>
+  </Provider>,
 );
