+++ src/spinner/spinner.types.ts
